refactor: extract DocumentEquivalenceComparer, IDocumentMerger, IDocumentLoader modules from OpenApiDocumentFactory
DocumentEquivalenceComparer: canonical JSON equivalence comparison
for merge conflict detection. Extracted from OpenApiDocumentFactory
with public methods for testability.

IDocumentMerger + DocumentMerger: multi-document merge logic
(paths, schemas, definitions, security schemes, tags).
Depends on DocumentEquivalenceComparer for conflict detection.

IDocumentLoader + DocumentLoader: document loading from file
or HTTP URL with YAML/JSON detection, OpenApiMultiFileReader
for external references, and NSwag fallback.

OpenApiDocumentFactory: deepened to thin composition of
IDocumentLoader + IDocumentMerger. Public API unchanged.
