+++ src/Refitter.Core/IDocumentLoader.cs
@@ -1,0 +1,9 @@
+using NSwag;
+using OpenApiDocument = NSwag.OpenApiDocument;
+
+namespace Refitter.Core;
+
+internal interface IDocumentLoader
+{
+    Task<OpenApiDocument> LoadAsync(string path);
+}
