+++ src/Refitter.Core/IDocumentMerger.cs
@@ -1,0 +1,9 @@
+using NSwag;
+using OpenApiDocument = NSwag.OpenApiDocument;
+
+namespace Refitter.Core;
+
+internal interface IDocumentMerger
+{
+    OpenApiDocument Merge(OpenApiDocument[] documents);
+}
